Cache機構をclassに分離した
